🚚 Rename: 컨벤션에 맞게 파일명 수정 및 폴더구조 변경
