+++ src/apis/onboarding-api.ts
@@ -1,5 +1,5 @@
 import { PostProfileRegistrationRequest, ProfileRegistrationResponse } from "@/types/onboarding";
-import { handleApiError } from "@/utils/api-error/error-handler";
+import { handleApiError } from "@/utils/api-error";
 import { apiClient } from ".";
 import { ENDPOINTS } from "./end-points";
