+++ src/apis/performance-api.ts
@@ -8,7 +8,7 @@ import {
   ServerFacilityInfo,
   ServerRecommendedSearchWords,
 } from "@/types/performance/server";
-import { handleApiError } from "@/utils/api-error/error-handler";
+import { handleApiError } from "@/utils/api-error";
 import { ENDPOINTS } from "./end-points";
 import { apiClient } from "./index";
