+++ src/components/search/add-info/result.tsx
@@ -3,15 +3,15 @@
 import Location from "@/assets/icons/location.svg";
 import Search from "@/assets/icons/search.svg";
 import Icon from "@/components/commons/icons";
+import LoadingSpinner from "@/components/commons/spinner";
+import Toast from "@/components/commons/toast";
 import { SectionLayout } from "@/components/layout";
 import { useSearchPerformances } from "@/hooks";
-import { useSearchStore } from "@/stores/useSearchStore";
-import useUploadTextStore from "@/stores/useUploadTextStore";
+import { useSearchStore } from "@/stores/use-search.store";
+import useUploadTextStore from "@/stores/use-upload-text.store";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import SearchCard from "../card";
-import LoadingSpinner from "@/components/commons/spinner";
-import Toast from "@/components/commons/toast";
 
 interface SelectPerformanceButtonProps {
   onConfirm: () => void;
