+++ src/stores/index.ts
@@ -1,2 +1,4 @@
 export * from "./use-onboarding.store";
+export * from "./use-search.store";
+export * from "./use-upload-text.store";
 export * from "./use-user.store";
