+++ src/stores/use-search.store.ts
