+++ src/styles/gradients/index.ts
@@ -1,0 +1,3 @@
+export { default as CommentGradient } from "./comment-gradient";
+export { default as LikeGradientDefault } from "./like-gradient-default";
+export { default as LikeGradientPressed } from "./like-gradient-pressed";
