+++ src/styles/gradients/like-gradient-default.tsx
