+++ src/styles/gradients/like-gradient-pressed.tsx
