+++ src/utils/api-error/index.ts
@@ -1,0 +1,1 @@
+export { handleApiError } from "./error-handler";
